[SIG-4375] Move StatusLayer component Since the `StatusLayer` component hasn't any dependency on the `WfsLayer` component, nor does it use its context, this PR moves the component to another folder. That folder, `VectorLayer` will be the context in which other related layer components will be placed.
