+++ src/signals/incident/components/form/MapSelectors/Asset/Selector/VectorLayer/StatusLayer/StatusLayer.test.tsx
