+++ src/signals/incident/components/form/MapSelectors/Asset/Selector/VectorLayer/StatusLayer/utils.ts
